+++ FreeRTOS_Source/tasks.c
@@ -273,8 +273,8 @@ typedef struct tskTaskControlBlock 			/* The old naming convention is used to pr
 
 	/*------------------------------  Extend to support validation: Start ------------------------------*/
 	#if ( configINTERMITTENT_DISTRIBUTED == 1)
-		uint64_t  vBegin;
-		uint64_t  vEnd;
+		uint32_t  vBegin;
+		uint32_t  vEnd;
 		/*------------------------------  Extend to support validation: End ------------------------------*/
 		/*------------------------------  Extend to support dynamic stack: Start -------------------------*/
 		void * AddressOfVMStack;
